pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gendiff"
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
readme = "README.md"
authors = [
    {name = "Konstantin Zabolotnev", email = "konstantin.zabolotnev@gmail.com"},
    {name = "Hexlet team", email = "team@hexlet.io"}
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.hatch.build]
include = [
    "gendiff/__init__.py",
    "gendiff/ast_builder.py",
    "gendiff/formatters/__init__.py",
    "gendiff/formatters/stylish.py",
    "gendiff/generate_diff.py",
    "gendiff/parsers.py",
    "gendiff/scripts/gendiff.py"
]
exclude = [
    "**/__pycache__",
    "**/*.pyc"
]

[tool.ruff]
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "Q",   # flake8-quotes
    "C90", # mccabe complexity
    "B",   # flake8-bugbear
]
ignore = ["E203"]
exclude = [
    "build",
    "dist",
    ".git",
    "__pycache__",
    ".venv"
]

[tool.ruff.isort]
known-first-party = ["gendiff"]
combine-as-imports = true

[tool.ruff.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"
docstring-quotes = "double"

[tool.ruff.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[tool.coverage.run]
source = ["gendiff"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:"
]
